// lib/api/model/processReceiptResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Pina API
 * OpenAPI spec version: 1.0
 */
import type { RecognizedItem } from "./recognizedItem";

export interface ProcessReceiptResponseDto {
  items: RecognizedItem[];
  /** Number of new items added to pantry */
  addedToPantry: number;
  /** Newly created pantry items */
  newItems: RecognizedItem[];
}